scripts/python/requirements.txt
aioschedule==0.5.2
aiosqlite==0.17.0
certifi==2021.10.8
chardet==4.0.0
gevent==21.8.0
greenlet==1.1.2
grequests==0.6.0
idna==3.3
importlib-metadata==4.8.2
loguru==0.5.3
python-dotenv==0.19.2
requests==2.26.0
schedule==1.1.0
SQLAlchemy==1.4.27
typing-extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
